Handle exceptions thrown "during" a suite with no children.
This is tricky to reproduce, but it can happen in certain async
scenarios in the presence of focused specs if an async spec tells
Jasmine that it's complete and then causes the global error event
to fire later on. In that situation we avoid crashing and record
the failure as a top level afterAll failure, so that it at least
get surfaced to the user somewhere.

Signed-off-by: Steve Gravrock <sgravrock@pivotal.io>